requirements.txt
Django == 2.2.10
django_rest_framework
psycopg2 >=2.8,<2.9